+++ src/HttpBasicAuthentication.php
@@ -125,7 +125,7 @@ final class HttpBasicAuthentication implements MiddlewareInterface
                     sprintf('Basic realm="%s"', $this->options["realm"])
                 );
 
-            return $this->processError($request, $response, [
+            return $this->processError($response, [
                 "message" => "Authentication failed"
             ]);
         }
@@ -190,13 +190,10 @@ final class HttpBasicAuthentication implements MiddlewareInterface
     /**
      * Execute the error handler.
      */
-    private function processError(
-        ServerRequestInterface $request,
-        ResponseInterface $response,
-        array $arguments
-    ): ResponseInterface {
+    private function processError(ResponseInterface $response, array $arguments): ResponseInterface
+    {
         if (is_callable($this->options["error"])) {
-            $handler_response = $this->options["error"]($request, $response, $arguments);
+            $handler_response = $this->options["error"]($response, $arguments);
             if ($handler_response instanceof ResponseInterface) {
                 return $handler_response;
             }